requirements.txt
python-dotenv
langchain
langchain-core
langchain-community
langchain-huggingface[community]
openai
groq
langchain-openai
langchain-groq
langchain-google-genai
langchain-mistralai
google-generativeai
langchain-huggingface
streamlit
langchain_openrouter
langchain-tavily